requirements.txt
pyTelegramBotAPI==4.14.0
openai>=1.0.0
python-dotenv==1.0.0
